Rename `polymorphic-component.ts` to `wordpress-component.ts`
