+++ packages/components/src/ui/context/context-connect.js
@@ -24,12 +24,12 @@ import { getStyledClassNameFromKey } from './get-styled-class-name-from-key';
  * The hope is that we can improve render performance by removing functional
  * component wrappers.
  *
- * @template {import('./polymorphic-component').WordPressComponentProps<{}, any, any>} P
+ * @template {import('./wordpress-component').WordPressComponentProps<{}, any, any>} P
  * @param {(props: P, ref: import('react').Ref<any>) => JSX.Element | null} Component            The component to register into the Context system.
  * @param {string}                                                          namespace            The namespace to register the component under.
  * @param {Object}                                                          options
  * @param {boolean}                                                         [options.memo=false]
- * @return {import('./polymorphic-component').WordPressComponentFromProps<P>} The connected WordPressComponent
+ * @return {import('./wordpress-component').WordPressComponentFromProps<P>} The connected WordPressComponent
  */
 export function contextConnect( Component, namespace, options = {} ) {
 	/* eslint-enable jsdoc/valid-types */
