+++ packages/components/src/ui/context/wordpress-component.ts
